fixed typing of the `get_win()` fn `win` param
`vim.api.nvim_get_current_win()` and `vim.api.nvim_list_wins()` return `integer` and `integer[]`, respectively. Integer seems to be the correct typing annotation here. Similarly, `edgebar.wins.win` is a window ID integer, so the final conditional of this function is comparing the `win` param to an integer.